Changed matrix indexing to 64bit

+++ src/gmc_io.c
@@ -20,8 +20,8 @@ matrix read_matrix(const char * path) {
     fscanf(fd, "%u %u", &w, &h);
     matrix m = new_matrix(w, h);
 
-    for(int y = 0; y < m.h; y++)
-        for(int x = 0; x < m.w; x++)
+    for(long long y = 0; y < m.h; y++)
+        for(long long x = 0; x < m.w; x++)
             fscanf(fd, "%lg", &m.data[y * m.w + x]);
 
     fclose(fd);
@@ -32,8 +32,8 @@ void dump_matrix(matrix m, const char * path) {
     FILE * fd = fopen(path, "w");
 
     fprintf(fd, "%u %u\n", m.w, m.h);
-    for(int y = 0; y < m.h; y++) {
-        for(int x = 0; x < m.w - 1; x++)
+    for(long long y = 0; y < m.h; y++) {
+        for(long long x = 0; x < m.w - 1; x++)
             fprintf(fd, "%.17lg ", m.data[y * m.w + x]);
         fprintf(fd, "%.17lg\n", m.data[y * m.w + m.w - 1]);
     }
@@ -45,10 +45,10 @@ void dump_sparse(sparse_matrix m, const char * path, int width) {
     matrix dense = new_matrix(width, m.h);
     memset(dense.data, 0x00, width * m.h * sizeof(double));
 
-    for(int y = 0; y < m.h; y++) 
+    for(long long y = 0; y < m.h; y++) 
         for(int i = 0; i < m.w; i++) {
             sprs_val val = m.data[y * m.w + i];
-            int x = val.i;
+            long long x = val.i;
             dense.data[y * width + x] = val.value;
         }
